Add solution to 1.45
The problem called for using the repeated function created from
problem 1.43. However, when I re-used my solution, I would start
to get a stack overflow in my tests on the 4th degree. I spent a
little time tryng to figure out why, but I am giving up for now.
Nonetheless, the spirit of the getFixedPointInput function is the
same.